Auto-media embed is now a separate plugin. Changed the delay from 500ms to 100ms.
